+++ src/MoBi.Presentation/Tasks/SimulationUpdateTask.cs
@@ -4,13 +4,9 @@ using MoBi.Core.Domain.Model;
 using MoBi.Core.Domain.Services;
 using MoBi.Core.Services;
 using MoBi.Presentation.Presenter;
-using MoBi.Presentation.Tasks.Interaction;
 using OSPSuite.Core.Commands.Core;
-using OSPSuite.Core.Domain;
 using OSPSuite.Core.Domain.Builder;
 using OSPSuite.Core.Domain.Services;
-using OSPSuite.Core.Extensions;
-using OSPSuite.Utility.Extensions;
 
 namespace MoBi.Presentation.Tasks
 {
@@ -26,62 +22,35 @@ namespace MoBi.Presentation.Tasks
       /// <returns></returns>
       ICommand ConfigureSimulation(IMoBiSimulation simulationToConfigure);
 
-      ICommand UpdateSimulation(IMoBiSimulation subject);
+      ICommand UpdateSimulation(IMoBiSimulation simulationToUpdate);
    }
 
    public class SimulationUpdateTask : ISimulationUpdateTask
    {
       private readonly IMoBiContext _context;
       private readonly IMoBiApplicationController _applicationController;
-      private readonly IEntityPathResolver _entityPathResolver;
       private readonly ISimulationFactory _simulationFactory;
       private readonly ICloneManagerForBuildingBlock _cloneManager;
-      private readonly IInteractionTasksForSimulation _interactionTasksForSimulation;
       private readonly ISimulationConfigurationFactory _simulationConfigurationFactory;
 
       public SimulationUpdateTask(IMoBiContext context,
          IMoBiApplicationController applicationController,
-         IEntityPathResolver entityPathResolver,
          ISimulationFactory simulationFactory,
          ICloneManagerForBuildingBlock cloneManager,
-         IInteractionTasksForSimulation interactionTasksForSimulation,
          ISimulationConfigurationFactory simulationConfigurationFactory)
       {
          _context = context;
          _applicationController = applicationController;
-         _entityPathResolver = entityPathResolver;
          _simulationFactory = simulationFactory;
          _cloneManager = cloneManager;
-         _interactionTasksForSimulation = interactionTasksForSimulation;
          _simulationConfigurationFactory = simulationConfigurationFactory;
       }
 
       public ICommand UpdateSimulation(IMoBiSimulation simulationToUpdate)
       {
-         var simulationConfiguration = _simulationConfigurationFactory.Create();
-         
-         simulationConfiguration.CopyPropertiesFrom(simulationToUpdate.Configuration);
-         simulationConfiguration.SimulationSettings = _cloneManager.Clone(simulationToUpdate.Configuration.SimulationSettings);
-         
-         simulationToUpdate.Configuration.ModuleConfigurations.Each(moduleConfiguration => { simulationConfiguration.AddModuleConfiguration(templateModuleConfigurationFor(moduleConfiguration)); });
+         var simulationConfiguration = _simulationConfigurationFactory.CreateFromProjectTemplatesBasedOn(simulationToUpdate.Configuration);
 
-         simulationConfiguration.Individual = templateBuildingBlockFor(simulationToUpdate.Configuration.Individual);
-
-         simulationToUpdate.Configuration.ExpressionProfiles.Each(x => { simulationConfiguration.AddExpressionProfile(templateBuildingBlockFor(x)); });
-
-         return updateSimulation(simulationToUpdate, simulationConfiguration, new MoBiMacroCommand(), AppConstants.Captions.UpdatingSimulation);
-      }
-
-      private ModuleConfiguration templateModuleConfigurationFor(ModuleConfiguration moduleConfiguration)
-      {
-         return new ModuleConfiguration(_interactionTasksForSimulation.TemplateModuleFor(moduleConfiguration.Module),
-            templateBuildingBlockFor(moduleConfiguration.SelectedInitialConditions),
-            templateBuildingBlockFor(moduleConfiguration.SelectedParameterValues));
-      }
-
-      private TBuildingBlock templateBuildingBlockFor<TBuildingBlock>(TBuildingBlock buildingBlock) where TBuildingBlock : class, IBuildingBlock
-      {
-         return _interactionTasksForSimulation.TemplateBuildingBlockFor(buildingBlock) as TBuildingBlock;
+         return updateSimulation(simulationToUpdate, simulationConfiguration, AppConstants.Captions.UpdatingSimulation);
       }
 
       public ICommand ConfigureSimulation(IMoBiSimulation simulationToConfigure)
@@ -95,13 +64,12 @@ namespace MoBi.Presentation.Tasks
          if (simulationConfiguration == null)
             return new MoBiEmptyCommand();
 
-         return updateSimulation(simulationToConfigure, simulationConfiguration, new MoBiMacroCommand());
+         return updateSimulation(simulationToConfigure, simulationConfiguration);
       }
 
       private ICommand<IMoBiContext> updateSimulation(
          IMoBiSimulation simulationToUpdate,
          SimulationConfiguration simulationConfigurationReferencingTemplates,
-         IMoBiCommand configurationCommands,
          string message = AppConstants.Captions.ConfiguringSimulation)
       {
          //create model using referencing templates
@@ -120,7 +88,6 @@ namespace MoBi.Presentation.Tasks
             ObjectType = updateSimulationCommand.ObjectType
          };
 
-         macro.Add(configurationCommands);
          macro.Add(updateSimulationCommand);
          return macro;
       }
